requirements/base.txt
django-environ==0.4.5
django-extensions==2.2.8
django-url-checks==0.2.0
Django==3.2.13
djangorestframework==3.12.4
ipython==7.34.0
pytz==2019.3
whitenoise==5.0.1
Pillow==8.1.2
djangorestframework-gis==0.17
django-cors-headers==3.7.0
drf-spectacular==0.22.1
djoser==2.1.0
djangorestframework-simplejwt==5.2.0